Factor out reasoning about strings in own strategy
