+++ key.core/src/main/java/de/uka/ilkd/key/strategy/JavaCardDLStrategy.java
@@ -5,16 +5,12 @@ package de.uka.ilkd.key.strategy;
 
 import java.util.concurrent.atomic.AtomicLong;
 
-import de.uka.ilkd.key.ldt.BooleanLDT;
-import de.uka.ilkd.key.ldt.CharListLDT;
 import de.uka.ilkd.key.ldt.HeapLDT;
 import de.uka.ilkd.key.ldt.IntegerLDT;
 import de.uka.ilkd.key.ldt.LocSetLDT;
-import de.uka.ilkd.key.ldt.SeqLDT;
 import de.uka.ilkd.key.logic.op.Equality;
 import de.uka.ilkd.key.logic.op.Junctor;
 import de.uka.ilkd.key.logic.op.Quantifier;
-import de.uka.ilkd.key.logic.op.SortDependingFunction;
 import de.uka.ilkd.key.proof.Goal;
 import de.uka.ilkd.key.proof.Proof;
 import de.uka.ilkd.key.rule.UseDependencyContractRule;
@@ -47,7 +43,6 @@ import org.key_project.prover.strategy.costbased.feature.instantiator.ChoicePoin
 import org.key_project.prover.strategy.costbased.termProjection.ProjectionToTerm;
 import org.key_project.prover.strategy.costbased.termfeature.IsNonRigidTermFeature;
 import org.key_project.prover.strategy.costbased.termfeature.OperatorClassTF;
-import org.key_project.prover.strategy.costbased.termfeature.TermFeature;
 
 import org.jspecify.annotations.NonNull;
 
@@ -279,8 +274,6 @@ public class JavaCardDLStrategy extends AbstractFeatureStrategy {
             add(longConst(-7000), ScaleFeature.createScaled(findDepthFeature, 10.0)));
         bindRuleSet(d, "update_join", -4600);
         bindRuleSet(d, "update_apply", -4500);
-
-        setUpStringNormalisation(d);
 
         setupSplitting(d);
 
@@ -556,94 +549,6 @@ public class JavaCardDLStrategy extends AbstractFeatureStrategy {
             // hide auxiliary equation after the skolem constatns have
             // been replaced by it's computed value
             add(isSelectSkolemConstantTerm("auxiliarySK"), longConst(-500)));
-    }
-
-    private void setUpStringNormalisation(RuleSetDispatchFeature d) {
-
-        // translates an integer into its string representation
-        bindRuleSet(d, "integerToString", -10000);
-
-        // do not convert char to int when inside a string function
-        // feature used to recognize if one is inside a string literal
-        final SeqLDT seqLDT = getServices().getTypeConverter().getSeqLDT();
-        final CharListLDT charListLDT = getServices().getTypeConverter().getCharListLDT();
-        final BooleanLDT booleanLDT = getServices().getTypeConverter().getBooleanLDT();
-
-
-        final TermFeature keepChar =
-            or(op(seqLDT.getSeqSingleton()), or(op(charListLDT.getClIndexOfChar()),
-                or(op(charListLDT.getClReplace()), op(charListLDT.getClLastIndexOfChar()))));
-
-        bindRuleSet(d, "charLiteral_to_intLiteral",
-            ifZero(isBelow(keepChar), inftyConst(), longConst(-100)));
-
-        // establish normalform
-
-        // tf below only for test
-        final TermFeature anyLiteral = or(tf.charLiteral,
-            or(tf.literal, op(booleanLDT.getFalseConst()), op(booleanLDT.getTrueConst())));
-
-        final TermFeature seqLiteral = rec(anyLiteral, or(op(seqLDT.getSeqConcat()),
-            or(op(seqLDT.getSeqSingleton()), or(anyLiteral, inftyTermConst()))));
-
-        Feature belowModOpPenality = ifZero(isBelow(ff.modalOperator), longConst(500));
-
-        bindRuleSet(d, "defOpsSeqEquality",
-            add(NonDuplicateAppModPositionFeature.INSTANCE,
-                ifZero(add(applyTF("left", seqLiteral), applyTF("right", seqLiteral)),
-                    longConst(1000), inftyConst()),
-                belowModOpPenality));
-
-        bindRuleSet(d, "defOpsConcat",
-            add(NonDuplicateAppModPositionFeature.INSTANCE,
-                ifZero(
-                    or(applyTF("leftStr", not(seqLiteral)), applyTF("rightStr", not(seqLiteral))),
-                    longConst(1000)
-                // concat is often introduced for construction purposes,
-                // we do not want to use its definition right at the
-                // beginning
-                ), belowModOpPenality));
-
-        bindRuleSet(d, "stringsSimplify", longConst(-5000));
-
-        final TermFeature charOrIntLiteral = or(tf.charLiteral, tf.literal,
-            or(add(OperatorClassTF.create(SortDependingFunction.class), // XXX:
-                // was CastFunctionSymbol.class
-                sub(tf.literal)), inftyTermConst()));
-
-        bindRuleSet(d, "defOpsReplaceInline",
-            ifZero(add(applyTF("str", seqLiteral), applyTF("searchChar", charOrIntLiteral),
-                applyTF("replChar", charOrIntLiteral)), longConst(-2500), inftyConst()));
-
-        bindRuleSet(d, "defOpsReplace", add(NonDuplicateAppModPositionFeature.INSTANCE,
-            ifZero(or(applyTF("str", not(seqLiteral)), applyTF("searchChar", not(charOrIntLiteral)),
-                applyTF("replChar", not(charOrIntLiteral))), longConst(500), inftyConst()),
-            belowModOpPenality));
-
-        bindRuleSet(d, "stringsReduceSubstring",
-            add(NonDuplicateAppModPositionFeature.INSTANCE, longConst(100)));
-
-        bindRuleSet(d, "defOpsStartsEndsWith", longConst(250));
-
-        bindRuleSet(d, "stringsConcatNotBothLiterals",
-            ifZero(MatchedAssumesFeature.INSTANCE, ifZero(
-                add(applyTF(instOf("leftStr"), seqLiteral),
-                    applyTF(instOf("rightStr"), seqLiteral)),
-                inftyConst()), inftyConst()));
-
-        bindRuleSet(d, "stringsReduceConcat", longConst(100));
-
-        bindRuleSet(d, "stringsReduceOrMoveOutsideConcat",
-            ifZero(NonDuplicateAppModPositionFeature.INSTANCE, longConst(800), inftyConst()));
-
-        bindRuleSet(d, "stringsMoveReplaceInside",
-            ifZero(NonDuplicateAppModPositionFeature.INSTANCE, longConst(400), inftyConst()));
-
-
-        bindRuleSet(d, "stringsExpandDefNormalOp", longConst(500));
-
-        bindRuleSet(d, "stringsContainsDefInline", SumFeature
-                .createSum(EqNonDuplicateAppFeature.INSTANCE, longConst(1000)));
     }
 
     private void setupReplaceKnown(RuleSetDispatchFeature d) {
